+++ src/main/java/acme/features/administrator/airline/AdministratorAirlineRepository.java
@@ -1,5 +1,5 @@
 
-package acme.entities.airline;
+package acme.features.administrator.airline;
 
 import java.util.Collection;
 
@@ -7,9 +7,10 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import acme.client.repositories.AbstractRepository;
+import acme.entities.airline.Airline;
 
 @Repository
-public interface AirlineRepository extends AbstractRepository {
+public interface AdministratorAirlineRepository extends AbstractRepository {
 
 	@Query("select a from Airline a")
 	Collection<Airline> findAllAirlines();
