+++ src/main/java/acme/features/administrator/airline/AdministratorAirlineUpdateService.java
@@ -1,5 +1,5 @@
 
-package acme.features.authenticated.administrator.airline;
+package acme.features.administrator.airline;
 
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -9,18 +9,17 @@ import acme.client.components.views.SelectChoices;
 import acme.client.services.AbstractGuiService;
 import acme.client.services.GuiService;
 import acme.entities.airline.Airline;
-import acme.entities.airline.AirlineRepository;
 import acme.entities.airline.AirlineType;
 
 @GuiService
-public class AdministratorAirlineCreateService extends AbstractGuiService<Administrator, Airline> {
+public class AdministratorAirlineUpdateService extends AbstractGuiService<Administrator, Airline> {
 
 	// Internal state ---------------------------------------------------------
 
 	@Autowired
-	private AirlineRepository repository;
+	private AdministratorAirlineRepository repository;
 
-	// AbstractGuiService interface -------------------------------------------
+	// AbstractGuiService interfaced ------------------------------------------
 
 
 	@Override
@@ -35,8 +34,10 @@ public class AdministratorAirlineCreateService extends AbstractGuiService<Admini
 	@Override
 	public void load() {
 		Airline airline;
+		int id;
 
-		airline = new Airline();
+		id = super.getRequest().getData("id", int.class);
+		airline = this.repository.findAirlineById(id);
 
 		super.getBuffer().addData(airline);
 	}
